backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.7
qdrant-client==1.8.0
psycopg[binary,pool]==3.1.17